feat: Expanded code structer, added Swagger and constructions
