+++ product-service/package.json
@@ -4,7 +4,8 @@
   "description": "Product service (API based)",
   "main": "handler.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "get-swagger-documentation": "serverless downloadDocumentation --outputFileName=swagger.ext"
   },
   "author": "",
   "license": "ISC",
@@ -14,6 +15,7 @@
     "babel-preset-es2015": "^6.24.1"
   },
   "devDependencies": {
+    "serverless-aws-documentation": "^1.1.0",
     "serverless-webpack": "^5.3.5",
     "webpack": "^4.44.1",
     "webpack-node-externals": "^2.5.0"
