+++ product-service/src/utils/product-filter.js
@@ -1,0 +1,8 @@
+export const  productFilter = (eventProductId, productList) => {
+
+  const resultProduct = productList.filter(product => {
+        return product.id === eventProductId;
+  })[0];
+  
+  return resultProduct;
+};
