+++ product-service/swagger.ext
@@ -1,0 +1,115 @@
+{
+  "swagger" : "2.0",
+  "info" : {
+    "description" : "Some example API",
+    "version" : "2020-11-01T23:26:49Z",
+    "title" : "dev-product-service",
+    "termsOfService" : "https://www.google.com",
+    "contact" : {
+      "name" : "Alexey Tomashevich",
+      "url" : "https://join.skype.com/invite/dS2VsXIdTG4z",
+      "email" : "lex.bogos@gmail.com"
+    },
+    "license" : {
+      "name" : "The license",
+      "url" : "https://www.github.com/lexBogos"
+    }
+  },
+  "host" : "z9f5582xt9.execute-api.eu-west-1.amazonaws.com",
+  "basePath" : "/dev",
+  "tags" : [ {
+    "name" : "Tag",
+    "description" : "First service tag"
+  } ],
+  "schemes" : [ "https" ],
+  "paths" : {
+    "/products" : {
+      "get" : {
+        "produces" : [ "application/json" ],
+        "responses" : {
+          "200" : {
+            "description" : "200 response",
+            "schema" : {
+              "$ref" : "#/definitions/getAllProductsResponse"
+            }
+          }
+        }
+      }
+    },
+    "/products/{productId}" : {
+      "get" : {
+        "produces" : [ "application/json" ],
+        "parameters" : [ {
+          "name" : "productId",
+          "in" : "path",
+          "required" : true,
+          "type" : "string"
+        } ],
+        "responses" : {
+          "200" : {
+            "description" : "200 response",
+            "schema" : {
+              "$ref" : "#/definitions/getProductByIdResponse"
+            }
+          }
+        }
+      }
+    }
+  },
+  "definitions" : {
+    "getProductByIdResponse" : {
+      "type" : "object",
+      "properties" : {
+        "price" : {
+          "type" : "string",
+          "description" : "Conditional price"
+        },
+        "count" : {
+          "type" : "number",
+          "description" : "The quantity of this product in the store"
+        },
+        "description" : {
+          "type" : "string",
+          "description" : "Product discription"
+        },
+        "id" : {
+          "type" : "string",
+          "description" : "Product id"
+        },
+        "title" : {
+          "type" : "string",
+          "description" : "Product title"
+        }
+      }
+    },
+    "getAllProductsResponse" : {
+      "type" : "array",
+      "description" : "An array of all products that are located in the store DB",
+      "items" : {
+        "type" : "object",
+        "properties" : {
+          "price" : {
+            "type" : "string",
+            "description" : "Conditional price"
+          },
+          "count" : {
+            "type" : "number",
+            "description" : "The quantity of this product in the store"
+          },
+          "description" : {
+            "type" : "string",
+            "description" : "Product discription"
+          },
+          "id" : {
+            "type" : "string",
+            "description" : "Product id"
+          },
+          "title" : {
+            "type" : "string",
+            "description" : "Product title"
+          }
+        }
+      }
+    }
+  }
+}
